timer ve kronometre için sayfa hazırlandı
